+++ packages/sdk/CHANGELOG.md
@@ -67,6 +67,6 @@
 - Rename spend transaction `value` to `valueToSpend`
 - Add `NocturneContext` object and add functionality for converting an asset request and desired operation to a `PostProofOperation` containing potentially several spend txs
 - Update `generateSpend2TestCase` script to write to `/fixtures`
-- Update `NocturneSigner` to derive `vk` from `sk` and use simplified 2 field `NocturneAddress`
+- Update `NocturneSigner` to derive `vk` from `sk` and use simplified 2 field `StealthAddress`
 - Move spend2 circuit prove/verify methods into sdk
 - Add babyjub classes and poseidon tree